backend/requirements.txt
# Core Backend
fastapi
uvicorn
pymongo
dnspython

# ML Dependencies
numpy
pandas
scikit-learn
joblib

#Utilities
requests
pydantic
python-multipart

#Deployment
locust

# security
python-dotenv
certifi

